Refactor/eval utils (#11)
* refactor: simplify GADM evaluation script with utility functions

* merge gadm_utils.py back into eval_gadm.py

* refactor: reorganize imports in eval_gadm.py script

* refactor: Reorganize module structure for improved readability and flow

* refactor: update import and reformat unittest_gadm_equality.py
